Moved StoreRecordHandler in handlers.store
